Add ML CoNLL-2012 partitions for the English coreference layer
There are two levels of partitions:

 i) Across the entire data
ii) Folds within the training partition

The partitions were created using the last digit in the file's
ID. Therefore there are a total of 8 folds in the training partition
and the development partition contains file IDs ending in 0 whereas
the ones in the test partition contain file IDs MOSTLY ending in 9.
MOSTLY, because some of the genre contained less than 10 files and
we tried to reserve the largest numerical file ID for the test partition.